Update LocalMediaProcessor.php
If the media files contain special characters or german umlauts, it may lead to problems storing the files in the file system. Example: file "groß.jpg" will be stored as "gro??.jpg" and subsequently the Thumbnail process will fail. The quickfix is to simply urlencode the filename. I also would suggest not to take the image description from magento as default file_name, why not take the original file name?